Move Api initalizer code into own class
